+++ database/archive/Generator/static/output/lightcurve.fits
